compile & commit built code in dist
code in src/ is now compiled into dist/ as ES6 javascript. because the
package.json points to this new directory's index and we commit it to
the repo, any project making use of this will maintain backwards
compatibility and will not need to download typescript and build the
code themselves.