[WR-114] rework RoleMapping -> GroupRoleMapping
the previous version of this class hinted that the object represented
something it did not. it is not a representation of all the roles
available at a realm, but rather the roles to which a particular group
maps to.

**MINOR BREAKING API CHANGE**
the previous class was publicly exposed by way of `DetailedIdentity.roles`
this commit contains a property name change `clients` -> `client` to match
Tozny's API response type. `DetailedIdentity` is returned by the client's
`identityDetails` method. as such, a migration is necessary if you made
use of `DetailedIdentity.role.clients`. the property is now `client`.